コントローラーの分離
